Adição do template pra monitorar os gestos + biblioteca de gestos e de animações (essa segunda ainda não foi usada). Agora arrastar pra direita abre o menu lateral
