+++ package.json
@@ -7,6 +7,7 @@
     "@fontsource/bebas-neue": "^4.5.9",
     "@fontsource/nunito": "^4.5.12",
     "@fontsource/quicksand": "^4.5.12",
+    "@react-spring/web": "^9.7.3",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
@@ -17,6 +18,7 @@
     "@types/react-dom": "^18.0.10",
     "@types/react-router-dom": "^5.3.3",
     "@types/styled-components": "^5.1.26",
+    "@use-gesture/react": "^10.3.0",
     "copy-to-clipboard": "^3.3.3",
     "gh-pages": "^4.0.0",
     "react": "^18.2.0",
